+++ src/Calibre/Filter.php
@@ -19,7 +19,7 @@ class Filter
 {
     public const PAGE_ID = PageId::FILTER_ID;
     public const PAGE_DETAIL = PageId::FILTER;
-    public const ROUTE_DETAIL = "?";
+    public const ROUTE_DETAIL = "page-filter";
     public const URL_PARAMS = [
         Author::URL_PARAM => Author::class,
         Language::URL_PARAM => Language::class,